Build : move bash-completion out of build process
skydive-bash-completion.sh is embedded inside skydive binary, via skydive
bash-completion command.
